Add Gallery component and update routing; remove old Gallery page
