+++ src/App.js
@@ -10,6 +10,7 @@ import FacialService from "./pages/service/FacialService.jsx";
 import ServicePage from "./pages/service/ServicePage.jsx";
 import Booking from "./pages/Booking.js";
 import ProductsPage from "./pages/ProductsPage";
+import Gallery from "./components/Gallery";
 
 function App() {
   return (
@@ -23,6 +24,7 @@ function App() {
           <Route path="/booking" element={<Booking />} />
           <Route path="/myAccount" element={<MyAccount />} />
           <Route path="/products/:clientLogin" element={<ProductsPage />} />
+          <Route path="/gallery" element={<Gallery />} />
         </Routes>
       </Layout>
     </Router>
