+++ src/pages/Home.css
@@ -381,59 +381,6 @@ p {
   margin: 0; /* Remove margens padrão para o parágrafo */
 }
 
-.our-gallery {
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  h2 {
-    margin: 80px 0;
-  }
-
-  .gallery {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: 2fr 1.5fr 2fr;
-    gap: 10px;
-    width: 80%;
-    max-width: 997px;
-    max-height: 978px;
-    margin-bottom: 80px;
-  }
-  .gallery img {
-    width: 100%;
-    height: auto;
-    display: block;
-  }
-
-  #img1 {
-    grid-column: 1 / 2;
-    grid-row: 1 / 3;
-  }
-  #img2 {
-    grid-column: 2 / 3;
-    grid-row: 1 / 2;
-  }
-  #img3 {
-    grid-column: 3 / 4;
-    grid-row: 1 / 4;
-  }
-  #img4 {
-    grid-column: 1 / 2;
-    grid-row: 3 / 4;
-  }
-  #img5 {
-    grid-column: 2 / 3;
-    grid-row: 2 / 4;
-  }
-  #img6 {
-    grid-column: 3 / 4;
-    grid-row: 3 / 3;
-  }
-}
-
 .contact {
   width: 100vw;
   background-color: var(--green-50);
@@ -839,166 +786,119 @@ p {
       margin: 40px 0;
     }
 
-    .gallery {
-      display: grid;
-      grid-template-columns: repeat(3, 1fr);
-      grid-template-rows: 2fr 1.5fr 2fr;
-      gap: 10px;
-      width: 80%;
-      max-width: 997px;
-      max-height: 978px;
-      margin-bottom: 40px;
-    }
-
-    .gallery img {
-      width: 100%;
-      height: auto;
-      display: block;
-    }
-
-    #img1 {
-      grid-column: 1 / 2;
-      grid-row: 1 / 3;
-    }
-
-    #img2 {
-      grid-column: 2 / 3;
-      grid-row: 1 / 2;
-    }
-
-    #img3 {
-      grid-column: 3 / 4;
-      grid-row: 1 / 4;
-    }
-
-    #img4 {
-      grid-column: 1 / 2;
-      grid-row: 3 / 4;
-    }
-
-    #img5 {
-      grid-column: 2 / 3;
-      grid-row: 2 / 4;
-    }
-
-    #img6 {
-      grid-column: 3 / 4;
-      grid-row: 3 / 3;
-    }
-  }
-
-  .contact {
-    overflow-x: hidden;
-    background-color: var(--green-50);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    box-sizing: border-box;
-
-    h2 {
-      margin: 40px 0 20px 0;
-      color: var(--white-200);
-    }
-
-    p {
-      color: var(--white-200);
-      font-size: 18px;
-      margin: 0 auto;
-      text-align: center;
-      max-width: 360px;
-    }
-
-    form {
+    .contact {
+      overflow-x: hidden;
+      background-color: var(--green-50);
       display: flex;
       flex-direction: column;
-      margin-top: 30px;
-      padding: 20px;
+      align-items: center;
       box-sizing: border-box;
-      width: 350px;
-      max-width: 400px;
 
-      .name-email {
+      h2 {
+        margin: 40px 0 20px 0;
+        color: var(--white-200);
+      }
+
+      p {
+        color: var(--white-200);
+        font-size: 18px;
+        margin: 0 auto;
+        text-align: center;
+        max-width: 360px;
+      }
+
+      form {
         display: flex;
         flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        gap: 24px;
-        width: 100%;
-        margin-bottom: 24px;
-        width: calc(100% - 20px);
-      }
-      .form-group,
-      .form-group-phone {
-        width: 100%;
-        margin-bottom: 20px;
-      }
-      .form-group-phone {
-        width: calc(100% - 20px);
-      }
-      .form-group input,
-      .form-group-phone input,
-      textarea {
-        width: 100%;
-        padding: 10px;
-        font-size: 1em;
-      }
+        margin-top: 30px;
+        padding: 20px;
+        box-sizing: border-box;
+        width: 350px;
+        max-width: 400px;
 
-      label {
-        color: var(--White-100, #fff);
-        font-family: Raleway;
-        font-size: 24px;
-        font-style: normal;
-        font-weight: 400;
-        line-height: normal;
-        letter-spacing: 4.8px;
-        margin-bottom: 5px;
-      }
+        .name-email {
+          display: flex;
+          flex-direction: column;
+          justify-content: center;
+          align-items: center;
+          gap: 24px;
+          width: 100%;
+          margin-bottom: 24px;
+          width: calc(100% - 20px);
+        }
+        .form-group,
+        .form-group-phone {
+          width: 100%;
+          margin-bottom: 20px;
+        }
+        .form-group-phone {
+          width: calc(100% - 20px);
+        }
+        .form-group input,
+        .form-group-phone input,
+        textarea {
+          width: 100%;
+          padding: 10px;
+          font-size: 1em;
+        }
 
-      input,
-      textarea {
-        border: 1px solid var(--white-200);
-        background-color: var(--green-50);
-        padding: 10px;
-        margin-right: 0;
-        font-size: 1em;
-        border: 1px solid #ccc;
-        color: var(--white-200);
-        border: 1px solid var(--white-200);
-        width: calc(100% - 20px);
-      }
+        label {
+          color: var(--White-100, #fff);
+          font-family: Raleway;
+          font-size: 24px;
+          font-style: normal;
+          font-weight: 400;
+          line-height: normal;
+          letter-spacing: 4.8px;
+          margin-bottom: 5px;
+        }
 
-      textarea {
-        height: 220px;
-      }
+        input,
+        textarea {
+          border: 1px solid var(--white-200);
+          background-color: var(--green-50);
+          padding: 10px;
+          margin-right: 0;
+          font-size: 1em;
+          border: 1px solid #ccc;
+          color: var(--white-200);
+          border: 1px solid var(--white-200);
+          width: calc(100% - 20px);
+        }
 
-      input::placeholder,
-      textarea::placeholder {
-        color: var(--white-200);
-        font-family: Raleway;
-        font-size: 14px;
-        font-style: normal;
-        font-weight: 400;
-        line-height: normal;
-        letter-spacing: 2.8px;
-      }
+        textarea {
+          height: 220px;
+        }
 
-      input:focus,
-      textarea:focus {
-        outline: none;
-        border-color: var(--white-200);
-        font-family: Raleway;
-      }
+        input::placeholder,
+        textarea::placeholder {
+          color: var(--white-200);
+          font-family: Raleway;
+          font-size: 14px;
+          font-style: normal;
+          font-weight: 400;
+          line-height: normal;
+          letter-spacing: 2.8px;
+        }
 
-      .btn-send {
-        display: flex;
-        justify-content: center;
-        margin: 30px 0 60px 0;
+        input:focus,
+        textarea:focus {
+          outline: none;
+          border-color: var(--white-200);
+          font-family: Raleway;
+        }
+
+        .btn-send {
+          display: flex;
+          justify-content: center;
+          margin: 30px 0 60px 0;
+        }
       }
     }
-  }
 
-  .blank {
-    height: 80px;
-    background-color: var(--white-200);
+    .blank {
+      height: 80px;
+      background-color: var(--white-200);
+    }
   }
 }
